most button errors

+++ ca.mcgill.ecse.climbsafe/app/src/main/java/ca/mcgill/ecse/climbsafe/view/controllers/UpdateMemberPageController.java
@@ -209,7 +209,7 @@ public class UpdateMemberPageController {
 	public void deleteMemberClicked(ActionEvent event) {
 		String email = memberChoiceBox.getValue();
 		boolean confirmDelete = deleteConfirmButton.isSelected();
-		if(confirmDelete) {
+		if(confirmDelete && email != null) {
 			try {
 				if(successful(() -> ClimbSafeFeatureSet1Controller.deleteMember(email))) {
 					memberChoiceBox.setValue(null);
